sow-api/pyproject.toml
[project]
name = "sow-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.10",
    "python-jose[cryptography]>=3.4.0",
    "python-keycloak>=5.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
